Basic map is displayed, see details
I don't like the current state of the map system. I wanted it to be
simple, but forcing the graphics into it has muddled it up. I need to
rethink the graphics.

The Tileset class is overly complicated. It was originally designed to
speed the program up, but it's too complicated to be of any use, and
almost unmaintainable. I shouldn't have to rely on the SurfaceManager
class for everything, so the Tileset should load it's own file.

Also, the associaction of the layers and the tilesets complicate things
even further. Each layer should have its own tileset, but to do this, I'll
need to also be able to reuse one tileset _between_ layers.

Finally, I need to ensure that the drawing routine is efficient,
especially when there are too many tiles to draw onto the screen at once.