The menu bar seems solid, and is responding correctly
What I've essentially done without realizing it is create drop down boxes,
and used those as part of the menu bar. This isn't a bad thing, but both
systems need to be separated and hardened. However, I'm happy with my
progress so far.

The next stage is to hook the buttons up to the program itself, which
shouldn't be too difficult.