Refactored MenuBar, it's now fully functional
This seems solid enough for general use, and may in fact be the first part
of several GUI modules. Although the menu bar will be a massive help in
the future, I still need several other modules before I can get back to
the editor proper. I've planned about a week ahead, so I know I just need
to work at it.

The menu bar still needs to be hooked up to the rest of the code, it's
fine as it is. I feel as though planning ahead might actually be a _good_
thing. Who knew?